backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Daytona Dashboard Backend"
authors = ["Devin <devin@devin.dev>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
